requirements.txt
asgiref==3.8.1
async-timeout==4.0.3
channels==4.1.0
channels-redis==4.2.0
Django==5.1.1
djangorestframework==3.15.2
drf-yasg==1.21.7
inflection==0.5.1
load-dotenv==0.1.0
msgpack==1.1.0
packaging==24.1
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
redis==5.0.8
sqlparse==0.5.1
typing_extensions==4.12.2
uritemplate==4.1.1
